+++ src/imio/directory/core/contents/contact/content.py
@@ -56,6 +56,7 @@ class ContactCroppingProvider(BaseCroppingProvider):
             return [
                 "portrait_affiche",
                 "paysage_affiche",
+                "carre_affiche",
             ]
         return []