+++ src/imio/directory/core/subscribers.py
@@ -59,7 +59,9 @@ def modified_contact(obj, event):
             geocode_object(obj)
         elif "ILeadImageBehavior.image" in d.attributes:
             # we need to remove cropping information of previous image
-            remove_cropping(obj, "image", ["portrait_affiche", "paysage_affiche"])
+            remove_cropping(
+                obj, "image", ["portrait_affiche", "paysage_affiche", "carre_affiche"]
+            )
 
 
 def modified_entity(obj, event):